[Feat] #464 - consolidate readshorcutviews
