+++ HappyAnding/HappyAnding/Extensions/View/View+Gesture.swift
@@ -61,3 +61,13 @@ extension View {
     }
     
 }
+
+struct SizePreferenceKey: PreferenceKey {
+    typealias Value = CGSize
+    static var defaultValue: Value = .zero
+    
+    static func reduce(value _: inout Value, nextValue: () -> Value) {
+        _ = nextValue()
+    }
+}
+
